+++ CHANGES.md
@@ -4,7 +4,7 @@ Change Log
 Beta Releases
 -------------
 
-### b20 - 2013-09-01
+### b20 - 2013-09-03
 
 * Breaking changes:
     * Replaced `ExtentGeometry` parameters for extruded extent to make them consistent with other geometries.
@@ -32,6 +32,7 @@ var geometry = BoxGeometry.createGeometry(box);
 
     * Removed `createTypedArray` and `createArrayBufferView` from each of the `ComponentDatatype` enumerations. Instead, use `ComponentDatatype.createTypedArray` and `ComponentDatatype.createArrayBufferView`.
     * `DataSourceDisplay` now requires a `DataSourceCollection` to be passed into its constructor.
+    * Replaced `createPickFragmentShaderSource` with `createShaderSource`.
 * Added outline geometries.  [#1021](https://github.com/AnalyticalGraphicsInc/cesium/pull/1021).
 * Added `EllipsoidGeodesic`.
 * Added `PolylinePipeline.scaleToSurface`.